fix: messages from server are show
